+++ src/main/java/group6/ui/helpers/builders/dialog/RemovalDialogBuilder.java
@@ -1,4 +1,4 @@
-package group6.ui.helpers;
+package group6.ui.helpers.builders.dialog;
 
 import java.util.List;
 import java.util.function.Consumer;
