+++ src/main/java/group6/ui/views/NodeTabView.java
@@ -6,11 +6,10 @@ import group6.entity.device.SensorType;
 import group6.entity.node.ControlPanel;
 import group6.ui.controllers.GuiController;
 import group6.ui.controllers.NodeDeviceService;
-import group6.ui.helpers.DeviceDialogBuilder;
-
-import group6.ui.helpers.NodeTabLayoutBuilder;
-import group6.ui.helpers.RemovalDialogBuilder;
-import group6.ui.helpers.UiAlerts;
+import group6.ui.helpers.builders.NodeTabLayoutBuilder;
+import group6.ui.helpers.builders.UiAlerts;
+import group6.ui.helpers.builders.dialog.DeviceDialogBuilder;
+import group6.ui.helpers.builders.dialog.RemovalDialogBuilder;
 import group6.protocol.DeviceKey;
 import group6.protocol.RefreshTarget;
 import javafx.scene.control.Label;
